chore: revert to commit 69f06876
