+++ apps/frontend/app/app/(app)/foodoffers/index.tsx
@@ -52,7 +52,6 @@ import BaseBottomSheet from '@/components/BaseBottomSheet';
 import type BottomSheet from '@gorhom/bottom-sheet';
 import CanteenSelectionSheet from '@/components/CanteenSelectionSheet/CanteenSelectionSheet';
 import SortSheet from '@/components/SortSheet/SortSheet';
-import CanteenSelection from "@/components/CanteenSelection";
 import HourSheet from '@/components/HoursSheet/HoursSheet';
 import CalendarSheet from '@/components/CalendarSheet/CalendarSheet';
 import { excerpt } from '@/constants/HelperFunctions';
@@ -167,14 +166,6 @@ const index: React.FC<DrawerContentComponentProps> = ({ navigation }) => {
 
   // Set Page Title
   useSetPageTitle(selectedCanteen?.alias || TranslationKeys.food_offers);
-
-  if (!selectedCanteen) {
-    return (
-      <SafeAreaView style={{ flex: 1, backgroundColor: theme.screen.background }}>
-        <CanteenSelection />
-      </SafeAreaView>
-    );
-  }
 
   useFocusEffect(
     useCallback(() => {
@@ -492,9 +483,9 @@ const index: React.FC<DrawerContentComponentProps> = ({ navigation }) => {
       setLoading(true);
       const canteenId = selectedCanteen?.id as string;
       let foodOffers =
-        prefetchedFoodOffers[canteenId]?.[selectedDate]
+        prefetchedFoodOffers[canteenId]?.[selectedDate];
 
-      if (!foodOffers && canteenId) {
+      if (!foodOffers) {
         const foodData = await fetchFoodOffersByCanteen(
           canteenId,
           selectedDate
@@ -676,12 +667,10 @@ const index: React.FC<DrawerContentComponentProps> = ({ navigation }) => {
                   activeOpacity={0.7}
                 >
                   <Text style={{ ...styles.heading, color: theme.header.text }}>
-                    {selectedCanteen
-                      ? excerpt(
-                          String(selectedCanteen.alias),
-                          screenWidth > 800 ? 30 : 10,
-                        )
-                      : translate(TranslationKeys.food_offers)}
+                    {excerpt(
+                      String(selectedCanteen?.alias),
+                      screenWidth > 800 ? 30 : 10
+                    ) || 'Food Offers'}
                   </Text>
                 </TouchableOpacity>
               </View>
