--- apps/frontend/app/components/CanteenSelection/index.ts
@@ -1,2 +1,0 @@
-export { default } from './CanteenSelection';
-export type { CanteenSelectionProps, CanteenProps } from './types';
