--- apps/frontend/app/components/CanteenSelection/types.ts
@@ -1,10 +1,0 @@
-import { DatabaseTypes } from 'repo-depkit-common';
-
-export interface CanteenSelectionProps {
-  closeSheet?: () => void;
-}
-
-export interface CanteenProps extends DatabaseTypes.Canteens {
-  imageAssetId: string;
-  image_url: string;
-}
